refactor(redux): remove side effects from predicates actions
